requirements.txt
gradio>=5.0.0, <6.0.0
numpy
opencv-python
Pillow
trimesh>=4.0.0
scipy
pystray
pytz